requirements.txt
aiohttp==3.7.4.post0
apispec==5.0.0
async-timeout==3.0.1
attrs==20.3.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
discord.py==1.7.1
Flask==2.0.1
flask-apispec==0.11.0
flask-marshmallow==0.14.0
idna==3.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
multidict==5.1.0
python-dotenv==0.17.1
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
webargs==8.0.0
Werkzeug==2.0.1
yarl==1.6.3
